Increase size of int and doubles

+++ clock.c
@@ -33,13 +33,13 @@ unsigned int compare_clocks(struct clock c1, struct clock c2) {
     return -1;
 }
 
-double clock_to_seconds(struct clock c) {
-    double seconds = c.seconds;
+long double clock_to_seconds(struct clock c) {
+    long double seconds = c.seconds;
     seconds += (c.nanoseconds / ONE_BILLION);
     return seconds;
 }
 
-struct clock seconds_to_clock(double seconds) {
+struct clock seconds_to_clock(long double seconds) {
     struct clock clk = { .seconds = (int)seconds };
     seconds -= clk.seconds;
     clk.nanoseconds = seconds * ONE_BILLION;
@@ -47,14 +47,14 @@ struct clock seconds_to_clock(double seconds) {
 }
 
 struct clock calculate_avg_time(struct clock clk, int divisor) {
-    double seconds = clock_to_seconds(clk);
-    double avg_seconds = seconds / divisor;
+    long double seconds = clock_to_seconds(clk);
+    long double avg_seconds = seconds / divisor;
     return seconds_to_clock(avg_seconds);
 }
 
 struct clock subtract_clocks(struct clock c1, struct clock c2) {
-    double seconds1 = clock_to_seconds(c1);
-    double seconds2 = clock_to_seconds(c1);
-    double result = seconds1 - seconds2;
+    long double seconds1 = clock_to_seconds(c1);
+    long double seconds2 = clock_to_seconds(c1);
+    long double result = seconds1 - seconds2;
     return seconds_to_clock(result);
 }
